+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Possibly fixed kilju bucket lid detaching when left behind
 - Further fixes regarding loading of rule files
 - Fixed exhaust muffler trigger sometimes not being enabled back when muffler detaches
+- Fixed duplicate rule files while using "mop rules" command
 
 ## 3.7.1 (29.06.2022)